backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "dishka>=1.10.0",
    "fastapi>=0.136.0",
    "redis>=7.4.0",
    "ruff>=0.15.11",
    "sqlalchemy>=2.0.49",
    "uuid6>=2025.0.1",
    "uvicorn>=0.44.0",
]
